# docs/openapi/user-api.yaml
openapi: 3.0.3
info:
  title: Meeting Agent API - User Management
  description: |
    User management and authentication endpoints for Meeting Agent API.

    Supports:
    - Email/password authentication flow
    - User profile management (CRUD)
    - User avatar management
    - Token refresh mechanism
  version: 1.0.0
  contact:
    name: Meeting Agent Team
    email: Annd11@fpt.com

servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server

tags:
  - name: Authentication
    description: Email/password auth and token management
  - name: User
    description: User profile and information management
  - name: User Avatar
    description: User avatar upload and retrieval

paths:
  # ============================================================================
  # Authentication Endpoints
  # ============================================================================

  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user with email and password
      description: Create a new user account and return authentication tokens.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: User registered successfully and tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate user with email and password
      description: Authenticate an existing user and return JWT access and refresh tokens.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange a refresh token for a new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: New access token generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefreshResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # User Profile Endpoints
  # ============================================================================

  /me:
    get:
      tags:
        - Authentication
      summary: Get current user information
      description: Retrieve the profile information of the currently authenticated user
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Authentication
      summary: Update current user information
      description: Update profile information of the currently authenticated user
      operationId: updateCurrentUser
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User information updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /users:
    get:
      tags:
        - User
      summary: Get users list with pagination and filtering
      description: Retrieve a paginated list of users with optional filtering and sorting
      operationId: getUsers
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Page number (starting from 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: order_by
          in: query
          required: false
          description: Field to order results by
          schema:
            type: string
            default: created_at
            example: "created_at"
        - name: dir
          in: query
          required: false
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: name
          in: query
          required: false
          description: Filter by user name (partial match - case insensitive)
          schema:
            type: string
        - name: email
          in: query
          required: false
          description: Filter by email address (partial match)
          schema:
            type: string
        - name: position
          in: query
          required: false
          description: Filter by job position
          schema:
            type: string
        - name: created_at_gte
          in: query
          required: false
          description: Filter users created on or after this date (ISO 8601 format)
          schema:
            type: string
            format: date-time
        - name: created_at_lte
          in: query
          required: false
          description: Filter users created on or before this date (ISO 8601 format)
          schema:
            type: string
            format: date-time
        - name: project_id
          in: query
          required: false
          description: Filter users by project membership
          schema:
            type: integer
      responses:
        '200':
          description: Users list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsersResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # User Avatar Endpoints
  # ============================================================================

  /users/avatar/upload:
    post:
      tags:
        - User Avatar
      summary: Upload user avatar
      description: |
        Upload a user avatar image with strict validation.
        Accepts PNG or JPEG formats only, maximum 1MB file size.
        Image must contain valid PNG or JPEG magic bytes.
      operationId: uploadUserAvatar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (PNG or JPEG, max 1MB)
      responses:
        '200':
          description: Avatar uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '413':
          description: Payload too large - file exceeds maximum size (1MB)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-codeSamples:
        - lang: curl
          source: |
            curl -X POST http://localhost:8000/api/v1/users/avatar/upload \
              -H "Authorization: Bearer $ACCESS_TOKEN" \
              -F "file=@avatar.jpg"

  /users/{user_id}/avatar:
    get:
      tags:
        - User Avatar
      summary: Get user avatar
      description: Retrieve a user's avatar image (public endpoint, no authentication required)
      operationId: getUserAvatar
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID
          schema:
            type: integer
      responses:
        '200':
          description: Avatar image retrieved successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # WebSocket Status Endpoints
  # ============================================================================

components:
  # ============================================================================
  # Security Schemes
  # ============================================================================

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token required for protected endpoints

  # ============================================================================
  # Schemas
  # ============================================================================

  schemas:
    # Common Response Schemas
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
          example: true
        message:
          type: string
          description: Human-readable message
          nullable: true
          example: "Operation successful"
        data:
          type: object
          description: Response data
          nullable: true
        errors:
          type: array
          items:
            type: string
          description: List of error messages (if any)
          nullable: true
        meta:
          type: object
          description: Additional metadata
          nullable: true

    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_prev
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total:
          type: integer
          description: Total number of items
          example: 150
        total_pages:
          type: integer
          description: Total number of pages
          example: 8
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false

    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        data:
          type: object
          nullable: true

    # Authentication Schemas
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: The refresh token to exchange for a new access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          description: User password
          example: "strongpassword123"

    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          description: User password
          example: "strongpassword123"
        name:
          type: string
          description: User's full name
          nullable: true
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          description: URL to user's avatar image
          nullable: true
          example: "https://example.com/avatars/user_1.jpg"
        bio:
          type: string
          description: User biography
          nullable: true
          example: "Software developer with 5 years experience"
        position:
          type: string
          description: User job position
          nullable: true
          example: "Senior Developer"

    TokenRefreshResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - access_token
                - token_type
                - expires_in
              properties:
                access_token:
                  type: string
                  description: New JWT access token
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                token_type:
                  type: string
                  fixed: "bearer"
                  example: "bearer"
                expires_in:
                  type: integer
                  description: Token expiration time in seconds
                  example: 3600

    AuthResponse:
      type: object
      required:
        - success
        - message
        - data
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Authentication successful"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/UserProfile'
            access_token:
              type: string
              description: JWT access token
            refresh_token:
              type: string
              description: Refresh token for obtaining new access tokens
            token_type:
              type: string
              fixed: "bearer"
            expires_in:
              type: integer
              description: Token expiration in seconds

    # User Schemas
    UserProfile:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          description: User ID
          example: 1
        email:
          type: string
          format: email
          description: User's email address (ASCII only)
          example: "user@example.com"
        name:
          type: string
          description: User's full name
          nullable: true
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          description: URL to user's avatar image
          nullable: true
          example: "https://example.com/avatars/user_1.jpg"
        bio:
          type: string
          description: User's biography
          nullable: true
          example: "Software developer with 5 years experience"
        position:
          type: string
          description: User's job position
          nullable: true
          example: "Senior Developer"
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
          nullable: true
          example: "2024-01-20T14:45:00Z"

    UserInfoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'

    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          description: User's full name
          nullable: true
          example: "John Doe"
        avatar_url:
          type: string
          format: uri
          description: URL to user's avatar image
          nullable: true
          example: "https://example.com/avatars/user_1.jpg"
        bio:
          type: string
          description: User's biography
          nullable: true
          example: "Software developer with 5 years experience"
        position:
          type: string
          description: User's job position
          nullable: true
          example: "Senior Developer"

    PaginatedUsersResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UserProfile'
            pagination:
              $ref: '#/components/schemas/PaginationMeta'

    # Avatar Upload Schemas
    AvatarUploadResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - avatar_url
              properties:
                avatar_url:
                  type: string
                  format: uri
                  description: URL of the uploaded avatar
                  example: "https://example.com/avatars/user_1.jpg"

  # ============================================================================
  # Responses
  # ============================================================================

  responses:
    BadRequestError:
      description: Bad request - Invalid input data or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Invalid input"
            data: null

    UnauthorizedError:
      description: Unauthorized - Invalid, missing, or expired authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "You are not logged in or your session has expired. Please log in again."
            data: null

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "User not found"
            data: null

    ValidationError:
      description: Validation error - Invalid input data format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Input data validation failed"
            data: null

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "An unexpected error occurred"
            data: null